reused encounter card in patient home and patient profile
